i/prompting: moved patterns to dedicated subpackage of prompting
