+++ interfaces/prompting/constraints.go
@@ -23,6 +23,7 @@ import (
 	"errors"
 	"fmt"
 
+	"github.com/snapcore/snapd/interfaces/prompting/patterns"
 	"github.com/snapcore/snapd/sandbox/apparmor/notify"
 	"github.com/snapcore/snapd/strutil"
 )
@@ -34,8 +35,8 @@ var (
 )
 
 type Constraints struct {
-	PathPattern *PathPattern `json:"path-pattern,omitempty"`
-	Permissions []string     `json:"permissions,omitempty"`
+	PathPattern *patterns.PathPattern `json:"path-pattern,omitempty"`
+	Permissions []string              `json:"permissions,omitempty"`
 }
 
 // ValidateForInterface returns nil if the constraints are valid for the given
@@ -86,7 +87,7 @@ func (c *Constraints) Match(path string) (bool, error) {
 	if c.PathPattern == nil {
 		return false, fmt.Errorf("invalid constraints: no path pattern")
 	}
-	match, err := PathPatternMatch(c.PathPattern.String(), path)
+	match, err := patterns.PathPatternMatch(c.PathPattern.String(), path)
 	if err != nil {
 		return false, fmt.Errorf("invalid constraints: %w", err)
 	}
